Update EscapinggLiterals.java
The numerical and text values in Java programs are known as "Literals". Literals are where double or sinlge quotes are required when typing string and that they are to be treated literally to avoid ant permanent termination of the string. This is done through the use of the "\" escape operator. For example: 

String SayWords =  "\"Say more words." She said angrily".

Escape sequences for Output Formatting:-

\n - Newline
\t - Tab
\b - Backspace
\r - Carriage return
\f - Formfeed
\\ - Backlash
\' - Single quote marks
\" - Double quote marks